+++ src/main/java/com/example/r1pple/model/User.java
@@ -19,7 +19,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id") // 显式映射列名
-    private Long userId;      // 使用驼峰命名
+    private Integer userId;      // 使用驼峰命名
 
     @CreatedDate
     @Column(name = "create_time", updatable = false)
