+++ src/main/java/com/example/r1pple/service/ArticleService.java
@@ -6,6 +6,6 @@ import com.example.r1pple.DTO.response.ArticleResponse;
 import java.util.List;
 
 public interface ArticleService {
-    ArticleResponse getArticleById(Long articleId);
+    ArticleResponse getArticleById(Integer articleId);
     List<ArticleListResponse> getAllArticles();
 }
